+++ lib/grape_oas/introspectors/entity_introspector_support/type_schema_resolver.rb
@@ -95,7 +95,9 @@ module GrapeOAS
           if defined?(Grape::Entity) && type <= Grape::Entity
             GrapeOAS.introspectors.build_schema(type, stack: @stack, registry: @registry)
           else
-            build_schema_for_primitive(type) || default_string_schema
+            build_schema_for_primitive(type) ||
+              GrapeOAS.type_resolvers.build_schema(type) ||
+              default_string_schema
           end
         end